move C++ profiler parts to their own header file
